BISERVER-8010: Implemented infocenter help link for time level format BISERVER-8010: Implemented infocenter help link for time level format.
The actual uri handler interface. Each application that uses the modeler should implement one that is appropriate for its ui framework (swt, gwt etc).
The modeler controller has a setter for an instance of such an uri handler.

Removed dependencies for the run target. This makes it easier to juggle with jars from other projects (metadata)